CNN-Classifier/requirements.txt
absl-py==0.2.2
astor==0.6.2
backports.functools-lru-cache==1.5
backports.weakref==1.0rc1
bleach==1.5.0
cycler==0.10.0
Django==1.11.14
enum34==1.1.6
funcsigs==1.0.2
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
html5lib==0.9999999
image==1.5.24
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
mock==2.0.0
numpy==1.14.5
opencv-python==3.4.1.15
pbr==4.0.4
Pillow==5.2.0
protobuf==3.6.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
scipy==1.1.0
six==1.11.0
subprocess32==3.5.2
tensorboard==1.8.0
tensorflow==1.2.1
termcolor==1.1.0
Werkzeug==0.14.1
